+++ apps/settings/js/panels/root/panel.js
@@ -4,20 +4,24 @@ define(function(require) {
   var SettingsPanel = require('modules/settings_panel');
   var Root = require('panels/root/root');
   var BatteryItem = require('panels/root/battery_item');
-  var StorageUSB = require('panels/root/storage_usb');
-  var StorageApp = require('panels/root/storage_app');
+  var StorageUSBItem = require('panels/root/storage_usb_item');
+  var StorageAppItem = require('panels/root/storage_app_item');
 
   return function ctor_root_panel() {
     var root = Root();
     var batteryItem;
 
-    var storage_usb = StorageUSB();
-    var storage_app = StorageApp();
+    var storageUsbItem = StorageUSBItem();
+    var storageAppItem = StorageAppItem();
 
     return SettingsPanel({
       onInit: function rp_onInit(panel) {
-        var storage_elements = {
-          appStorageDesc: panel.querySelector('.application-storage-desc'),
+        root.init();
+        batteryItem = BatteryItem(panel.querySelector('.battery-desc'));
+        storageUsbItem.init({
+          appStorageDesc: panel.querySelector('.application-storage-desc')
+        });
+        storageAppItem.init({
           mediaStorageDesc: panel.querySelector('.media-storage-desc'),
           umsEnabledCheckBox: panel.querySelector('.ums-switch-root'),
           umsEnabledInfoBlock: panel.querySelector('.ums-desc-root'),
@@ -25,17 +29,12 @@ define(function(require) {
           umsConfirmButton: panel.querySelector('.ums-confirm-option'),
           umsCancelButton: panel.querySelector('.ums-cancel-option'),
           mediaStorageSection: panel.querySelector('.media-storage-section')
-        };
-
-        root.init();
-        storage_usb.init(storage_elements);
-        storage_app.init(storage_elements);
-        batteryItem = BatteryItem(panel.querySelector('.battery-desc'));
+        });
       },
       onBeforeShow: function rp_onBeforeShow() {
         batteryItem.enabled = true;
-        storage_usb.enabled = true;
-        storage_app.enabled = true;
+        storageUsbItem.enabled = true;
+        storageAppItem.enabled = true;
       },
       onShow: function rp_onShow() {
         // XXX: Set data-ready to true to indicate that the first panel is
@@ -46,8 +45,8 @@ define(function(require) {
       },
       onHide: function rp_onHide() {
         batteryItem.enabled = false;
-        storage_usb.enabled = false;
-        storage_app.enabled = false;
+        storageUsbItem.enabled = false;
+        storageAppItem.enabled = false;
       }
     });
   };
