+++ apps/settings/js/panels/root/storage_app_item.js
