+++ apps/settings/js/panels/root/storage_usb_item.js
